backend/requirements.txt
#Base
Pillow==5.3.0

#Django
Django==2.2.4

#Django REST framework
djangorestframework==3.9.4
django-filter==2.0.0

#CORS
django-cors-headers==3.0.2

#JWT
djangorestframework-jwt==1.11.0

#ONLY DEV
#Tools
django-extensions==2.1.4